+++ addOns/reports/CHANGELOG.md
@@ -4,14 +4,16 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
-# Changed
+### Added
+- SARIF reporting
+
+### Changed
 - The XML and JSON reports now include programName metadata elements (Issue 6640).
 
 ## [0.16.0] - 2022-10-27
 ### Added
 - "XML Plus" report format for XML with requests and responses
 - Tags to "JSON Plus" report.
-- SARIF reporting
 
 ### Changed
 - Update minimum ZAP version to 2.12.0.